# server/src/middleware/userAuth.js
const jwt = require('jsonwebtoken')
const User = require('../models/user')
const auth = async (req, res, next) => {
    try {
        const token = req.header('Authorization').replace('Bearer ', '')
        const decoded = jwt.verify(token, 'example')
        const user = await User.findOne({
            _id: decoded._id,
            'tokens.token': token
        })
        if (!user) {
            throw new Error()

        }
        req.user = user
        req.token = token
        next()
    } catch {
        res.status(401).send({
            error: 'Please authenticate'
        })
    }
}
module.exports = auth