# server/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "CRUD",
    "description": "CRUD API"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "tags": [
    {
      "name": "Admins",
      "description": "API for Admin in the system"
    },
    {
      "name": "Users",
      "description": "API for the users"
    }
  ],
  "schemes": [
    "http"
  ],
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "For accessing the API a valid JWT token must be placed in all the queries in the 'Authorization' header. A valid token is generated by the API returned as answer of a call to the route /admin after creating an admin or /admin/login after giving a valid username & password. The following syntax must be used in the Authorization header: Bearer aaaaa.bbbbbb.cccccccc"
    }
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/admin": {
      "post": {
        "tags": [
          "Admins"
        ],
        "summary": "Create a new admin",
        "parameters": [
          {
            "name": "admin",
            "in": "body",
            "description": "Admin that we want to create",
            "schema": {
              "$ref": "#/definitions/Admin"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/AdminResponse"
            }
          },
          "400": {
            "description": "Error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/admin/login": {
      "post": {
        "tags": [
          "Admins"
        ],
        "summary": "Admin Login",
        "parameters": [
          {
            "name": "admin",
            "in": "body",
            "description": "Admins username and password",
            "schema": {
              "$ref": "#/definitions/AdminLogin"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/AdminResponse"
            }
          },
          "400": {
            "description": "Error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/admin/me": {
      "patch": {
        "tags": [
          "Admins"
        ],
        "summary": "Update admin",
        "security": [
          {
            "Bearer": []
          }
        ],
        "parameters": [
          {
            "name": "admin",
            "in": "body",
            "description": "Admin updates",
            "schema": {
              "$ref": "#/definitions/Admin"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/AdminResponse"
            }
          },
          "400": {
            "description": "Error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/admin/createUser": {
      "post": {
        "tags": [
          "Admins"
        ],
        "summary": "Create User",
        "security": [
          {
            "Bearer": []
          }
        ],
        "parameters": [
          {
            "name": "user",
            "in": "body",
            "description": "User Details",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/UserResponse"
            }
          },
          "400": {
            "description": "Error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/admin/listUsers": {
      "get": {
        "tags": [
          "Admins"
        ],
        "summary": "List all Users",
        "security": [
          {
            "Bearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/UserResponse"
            }
          },
          "400": {
            "description": "Error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/admin/updateUser/{id}": {
      "patch": {
        "tags": [
          "Admins"
        ],
        "summary": "Update user",
        "security": [
          {
            "Bearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the user",
            "required": true,
            "type": "string"
          },
          {
            "name": "user",
            "in": "body",
            "description": "Admin updates",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/AdminResponse"
            }
          },
          "400": {
            "description": "Error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/admin/logout": {
      "post": {
        "tags": [
          "Admins"
        ],
        "summary": "Admin Logout",
        "security": [
          {
            "Bearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/admin/logoutAll": {
      "post": {
        "tags": [
          "Admins"
        ],
        "summary": "Admin Logout from all devices",
        "security": [
          {
            "Bearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/user/login": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "User Login",
        "parameters": [
          {
            "name": "user",
            "in": "body",
            "description": "Users username and password",
            "schema": {
              "$ref": "#/definitions/UserLogin"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/UserLoginResponse"
            }
          },
          "400": {
            "description": "Error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/user/logout": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "User Logout",
        "security": [
          {
            "Bearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/user/logoutAll": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "User Logout from all devices",
        "security": [
          {
            "Bearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "Admin": {
      "required": [
        "firstName",
        "lastName",
        "email",
        "password",
        "mobileNo",
        "username",
        "accountType"
      ],
      "properties": {
        "firstName": {
          "type": "string",
          "example": "Ujjwal"
        },
        "lastName": {
          "type": "string",
          "example": "Kamilya"
        },
        "email": {
          "type": "string",
          "example": "ujjwal@gmail.com"
        },
        "mobileNo": {
          "type": "string",
          "example": "9932240669"
        },
        "username": {
          "type": "string",
          "example": "ujjwal"
        },
        "password": {
          "type": "string",
          "example": "1234567"
        }
      }
    },
    "User": {
      "required": [
        "firstName",
        "lastName",
        "email",
        "password",
        "mobileNo",
        "username",
        "accountType"
      ],
      "properties": {
        "firstName": {
          "type": "string",
          "example": "Ujjwal"
        },
        "lastName": {
          "type": "string",
          "example": "Kamilya"
        },
        "email": {
          "type": "string",
          "example": "ujjwalkamilya@gmail.com"
        },
        "mobileNo": {
          "type": "string",
          "example": "9932240669"
        },
        "username": {
          "type": "string",
          "example": "ujjwal2"
        },
        "password": {
          "type": "string",
          "example": "1234567"
        }
      }
    },
    "AdminResponse": {
      "properties": {
        "admin": {
          "properties": {
            "firstName": {
              "type": "string"
            },
            "lastName": {
              "type": "string"
            },
            "email": {
              "type": "string"
            },
            "mobileNo": {
              "type": "string"
            },
            "username": {
              "type": "string"
            },
            "password": {
              "type": "string"
            },
            "_id": {
              "type": "string"
            },
            "tokens": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "createdAt": {
              "type": "date",
              "example": "year-month-day-time"
            },
            "updatedAt": {
              "type": "date",
              "example": "year-month-day-time"
            }
          }
        },
        "token": {
          "type": "string"
        }
      }
    },
    "UserResponse": {
      "properties": {
        "firstName": {
          "type": "string"
        },
        "lastName": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "mobileNo": {
          "type": "string"
        },
        "username": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "_id": {
          "type": "string"
        },
        "tokens": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "createdAt": {
          "type": "date",
          "example": "year-month-day-time"
        },
        "updatedAt": {
          "type": "date",
          "example": "year-month-day-time"
        },
        "lastLogin": {
          "type": "date",
          "example": "year-month-day-time"
        }
      }
    },
    "Error": {
      "properties": {
        "error": {
          "type": "string"
        }
      }
    },
    "AdminLogin": {
      "required": [
        "username",
        "password"
      ],
      "properties": {
        "username": {
          "type": "string",
          "example": "ujjwal"
        },
        "password": {
          "type": "string",
          "example": "1234567"
        }
      }
    },
    "UserLogin": {
      "required": [
        "username",
        "password"
      ],
      "properties": {
        "username": {
          "type": "string",
          "example": "ujjwal2"
        },
        "password": {
          "type": "string",
          "example": "1234567"
        }
      }
    },
    "UserLoginResponse": {
      "properties": {
        "user": {
          "properties": {
            "firstName": {
              "type": "string"
            },
            "lastName": {
              "type": "string"
            },
            "email": {
              "type": "string"
            },
            "mobileNo": {
              "type": "string"
            },
            "username": {
              "type": "string"
            },
            "password": {
              "type": "string"
            },
            "_id": {
              "type": "string"
            },
            "tokens": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "createdAt": {
              "type": "date",
              "example": "year-month-day-time"
            },
            "updatedAt": {
              "type": "date",
              "example": "year-month-day-time"
            }
          }
        },
        "token": {
          "type": "string"
        }
      }
    }
  }
}